Add testutils module for shared unittesting code
