+++ tests/test_modules.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python2
 # -*- coding: utf-8-*-
 import unittest
-from client import test_mic, diagnose, jasperpath
+from client import testutils, diagnose, jasperpath
 from client.modules import Life, Joke, Time, Gmail, HN, News, Weather
 
 DEFAULT_PROFILE = {
@@ -29,7 +29,7 @@ class TestModules(unittest.TestCase):
         The server's responses, in a list.
         """
         self.assertTrue(module.is_valid(query))
-        mic = test_mic.Mic(inputs)
+        mic = testutils.TestMic(inputs)
         module.handle(query, mic, self.profile)
         return mic.outputs
